beanfiles/requirements.txt
absl-py==0.8.0
astor==0.8.0
boto3==1.12.39
botocore==1.15.39
certifi==2018.11.29
click==7.1.1
cycler==0.10.0
dnspython==1.16.0
docutils==0.15.2
Flask==1.1.2
gast==0.3.2
get==2019.4.13
google-pasta==0.1.7
grpcio==1.23.0
h5py==2.10.0
itsdangerous==1.1.0
Jinja2==2.11.1
jmespath==0.9.5
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.2.0
Markdown==3.1.1
MarkupSafe==1.1.1
matplotlib==3.2.1
mysql-connector-python==8.0.19
numpy==1.17.2
package-name==0.1
post==2019.4.13
protobuf==3.6.1
public==2019.4.13
pyparsing==2.4.7
python-dateutil==2.8.1
query-string==2019.4.13
request==2019.4.13
s3transfer==0.3.3
six==1.12.0
tb-nightly==1.14.0a20190603
tensorflow==2.0.0b0
termcolor==1.1.0
tf-estimator-nightly==1.14.0.dev2019060501
urllib3==1.25.8
virtualenv==16.7.5
Werkzeug==0.15.6
wrapt==1.11.2
